refactor: rename package from com.music4music to com.akitain
