+++ gradle.properties
@@ -8,5 +8,5 @@ fabric_version=0.141.3+1.21.11
 
 # Mod
 mod_version=0.1.0
-maven_group=com.music4music
+maven_group=com.akitain
 archives_base_name=enchantment-overhaul
