+++ src/main/resources/enchantment-overhaul.mixins.json
@@ -1,6 +1,6 @@
 {
     "required": true,
-    "package": "com.music4music.enchantmentoverhaul.mixin",
+    "package": "com.akitain.enchantmentoverhaul.mixin",
     "compatibilityLevel": "JAVA_21",
     "mixins": [
         "EnchantmentScreenHandlerMixin",
